Move InvalidAuth handling out of sensor
